fix: modify example usage of `@Model`
Modify example usage of the `@Model` annotation. The implementation in d09e8c6c did work, but not because of the `@Model` annotation.

This new implementation makes the current input-value accessible in `propValue` while the external property is called `input`. To demonstrate this the current value is always repeated below the input.

Modify `.eslintrc.js` to load new vue3 rules to allow `v-model` bindings with property names.